football-data/requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
ratelimit>=2.2.1
backoff>=2.2.1
tqdm>=4.66.0
